+++ airline-core/src/main/java/com/github/rvesse/airline/restrictions/options/MutuallyExclusiveRestriction.java
@@ -15,14 +15,12 @@
  */
 package com.github.rvesse.airline.restrictions.options;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 
 import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.collections4.Transformer;
 import org.apache.commons.lang3.tuple.Pair;
 
 import com.github.rvesse.airline.help.sections.HelpFormat;
@@ -33,8 +31,8 @@ import com.github.rvesse.airline.parser.errors.ParseOptionGroupException;
 import com.github.rvesse.airline.restrictions.OptionRestriction;
 import com.github.rvesse.airline.utils.predicates.parser.ParsedOptionFinder;
 import com.github.rvesse.airline.utils.predicates.restrictions.MutuallyExclusiveWithFinder;
+import com.github.rvesse.airline.utils.predicates.restrictions.MutuallyExclusiveWithOptionFinder;
 import com.github.rvesse.airline.utils.predicates.restrictions.MutuallyExclusiveWithTagParsedOptionFinder;
-import com.github.rvesse.airline.utils.predicates.restrictions.RequiredTagOptionFinder;
 
 public class MutuallyExclusiveRestriction implements OptionRestriction, HelpHint {
 
@@ -64,7 +62,7 @@ public class MutuallyExclusiveRestriction implements OptionRestriction, HelpHint
 
             // Otherwise may need to error
             if (parsedOptions.size() > 0 && otherParsedOptions.size() > parsedOptions.size()) {
-                Collection<OptionMetadata> taggedOptions = selectMetadata(otherParsedOptions);
+                Collection<OptionMetadata> taggedOptions = getTaggedOptions(state);
                 throw new ParseOptionGroupException(
                         "Only one of the following options may be specified but %d were found: %s", tag, taggedOptions,
                         otherParsedOptions.size(), toOptionsList(taggedOptions));
@@ -88,14 +86,16 @@ public class MutuallyExclusiveRestriction implements OptionRestriction, HelpHint
         return builder.toString();
     }
     
-    private Collection<OptionMetadata> selectMetadata(Collection<Pair<OptionMetadata, Object>> pairs){
-	ArrayList<OptionMetadata> metadata = new ArrayList<>();
-	for(Pair<OptionMetadata, Object> pair : pairs){
-	    metadata.add(pair.getLeft());
-	}
-	return metadata;
+    private <T> Collection<OptionMetadata> getTaggedOptions(ParseState<T> state) {
+        List<OptionMetadata> options = state.getCommand() != null ? state.getCommand().getAllOptions() : null;
+        if (options == null)
+            options = state.getGroup() != null ? state.getGroup().getOptions() : null;
+        if (options == null)
+            options = state.getGlobal() != null ? state.getGlobal().getOptions()
+                    : Collections.<OptionMetadata> emptyList();
+        return CollectionUtils.select(options, new MutuallyExclusiveWithOptionFinder(this.tag));
     }
-
+    
     public String getTag() {
         return tag;
     }